Refactor And Reorganize Page Menu Options
