+++ packages/bodiless-page/src/MenuOptions/index.ts
@@ -1,4 +1,3 @@
-export * from './withClonePageButton';
-export * from './withDeletePageButton';
-export * from './withMovePageButton';
-export * from './withNewPageButton';
+export * from './types';
+export * from './usePageMenuOptions';
+export * from './withPageMenuOptions';
