+++ packages/bodiless-page/src/MenuOptions/types.ts
@@ -1,0 +1,12 @@
+type PageMenuOptions = {
+  name: string,
+  icon: string,
+  label: string,
+  handler: any,
+  isDisabled: boolean,
+  isHidden: boolean,
+};
+
+export {
+  PageMenuOptions,
+};
