--- packages/bodiless-page/src/MenuOptions/withDeletePageButton.ts
@@ -1,57 +1,0 @@
-/**
- * Copyright © 2022 Johnson & Johnson
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { useCallback } from 'react';
-import {
-  ContextSubMenu,
-  BodilessBackendClient,
-  useEditContext,
-  withMenuOptions,
-} from '@bodiless/core';
-import {
-  menuFormPageDelete,
-} from '../Forms';
-
-const defaultClient = new BodilessBackendClient();
-
-const useMenuOptions = () => {
-  const context = useEditContext();
-  const menuOptions = [
-    {
-      name: 'page-group',
-      icon: 'description',
-      label: 'Page',
-      Component: ContextSubMenu,
-    },
-    {
-      name: 'deletepage',
-      icon: 'delete',
-      label: 'Delete',
-      group: 'page-group',
-      isHidden: useCallback(() => !context.isEdit, []),
-      handler: () => menuFormPageDelete(defaultClient),
-    },
-  ];
-  return menuOptions;
-};
-
-const withDeletePageButton = withMenuOptions({
-  useMenuOptions,
-  name: 'DeletePage',
-  root: true,
-});
-
-export {
-  withDeletePageButton,
-};
